[SKYLINE][SQL][PYTHON] Fix error where distinct in skyline queries was defined per dimension and not per skyline operator.
### What changes were proposed?

SQL query syntax, DataFrame, and DataSet API were changed for Scala/Java and Pyspark such that distinct is now specified per skyline operator instead of per dimension.

The nodes of the logical and physical plan were changed to accommodate this change. Distinctiveness is now stored in the skyline operator node itself instead of in the sequence of dimensions.

This included a rewrite of the internal logic of the skyline part of the DataFrame and DataSet API as well as changes to the query parsing.

### Why are the changes needed?

Distinct specification per dimension is not correct according to the definition of skyline queries found in [Stephan Börzsönyi, Donald Kossmann, Konrad Stocker: The Skyline Operator. ICDE 2001: 421-430]. Instead, distinct should be specified per skyline operator.

### Are there _any_ user-facing changes?

Yes, query string now follows the pattern of:
[ SKYLINE OF [ DISTINCT ] { expression { MIN | MAX | DIFF } [ , ... ] } ].

Additionally, the DataFrame and DataSet APIs of both Scala/Java and PySpark where changed in such a way that distinct is now specified via df.skylineDistinct(...) instead of via parameters. All instances where distinct was specified via a parameter were removed. This change introduced the consequence that .sdistinct() was removed in its entirety as well.