[SKYLINE][SQL] Implement first version of Block-Nested-Loop skyline algorithm as well as a basic distributed two-stage approach using the Block-Nested-Loop algorithm.
The distributed approach is based on the generic Map-Reduce approach of many local distributed skylines followed by a single global skyline.
Distribution of data is left to Spark for the local skyline and uses AllTuples for the global skyline.
For checking dominance, a generic utility is added which works with all relevant numeric data types.
Single-dimensional skylines are now optimized to use aggregates and filters instead of full skyline computations to increase performance and reduce runtime ideally to O(n) depending on the input.

### What changes were proposed?

Basic skyline implementation is added:
- Implementation of Block-Nested-Loop
- Implementation of basic distributed computation via local and global skyline (currently based on Block-Nested-Loop for both local and global skyline)
- Optimization of one-dimensional skylines by transforming them to a combination of aggregates and filters.

### Why are the changes needed?

Changes are needed to make basic skyline functionality available in Spark (SQL). They were necessary to make the Block-Nested-Loop algorithm and basic distribution functional. At the same time, single-dimensional skylines are now optimized.

### Are there _any_ user-facing changes?

The only user-facing change is that the skyline computations now should return proper results instead of acting as an identity function.
NOTE: Skyline behaviour may not be as expected on nullable columns/dimensions since the specialized algorithm to handle these cases was not implemented yet.