[SKYLINE][SQL] Implement incomplete version of distributed Block-Nested-Loop skyline algorithm as two-stage approach.
Implement detection of nullable columns such that complete or incomplete algorithms for skyline computation can be chosen automatically. Provide an override in SQL syntax as well as DataFrame and DataSet API to force complete skyline computation mode on skylines with nullable columns.

### What changes were proposed?

Implementation of incomplete skyline algorithm:
- Based on existing Block-Nested-Loop algorithm
- Implement detection of nullable columns where complete mode may not provide correct results
  -- chose incomplete algorithm for nullable columns automatically
  -- Provide override to force complete mode

### Why are the changes needed?

Changes are needed to allow computation of skylines on incomplete datasets, i.e., datasets where values are missing. Null values are treated as missing values in case of Spark.

### Are there _any_ user-facing changes?

User facing changes involve the overrides which allow the user to use complete algorithms even if the columns are nullable. Forcing the complete mode on nullable columns will only produce the expected results as long as no null value is encountered. If the complete mode is not forced, Spark will choose the incomplete algorithms if any nullable columns are encountered and the complete algorithms otherwise.

The changed skyline syntax introduces the new (optional) "COMPLETE" keyword and is as follows:
[ COMPLETE ] SKYLINE OF [ DISTINCT ] { expression { MIN | MAX | DIFF } [ , ... ] }

For the DataFrame and DataSet APIs we have already introduced the following skyline functions:
- skyline(...)
- skylineDistinct(...)
Additionally, we introduce new functions which behave exactly the same except forcing Spark to choose a complete skyline algorithm:
- skylineComplete(...)
- skylineDistinctComplete(...)