[SKYLINE][SQL] Extend Spark SQL analyzer such that skyline queries now work fully with aggregates and HAVING clauses.
Extend the analyzer such that:
- resolve aggregate attributes introduced by the aggregates
- resolve aggregates which are not part of the SELECT-FROM projection
- prevent premature projection in case of skylines and sorting/ordering
- re-add projection AFTER skyline or sorting/ordering

Slightly refactor code and rename the attributes of the skyline.

Add test cases to test the implementation of the skyline.

Slightly modify global computation of incomplete skylines to fix a bug.

Fix bug where using a complete and distinct skyline may result in a parser error by slightly changing the syntax of the skyline query.

### What changes were proposed?

Make skyline queries compatible with aggregates and HAVING.
Add test cases for basic testing of the implementation.

### Why are the changes needed?

Changes are needed to allow queries which contain a HAVING clause to also make use of skylines and sorting/ordering.
Add test cases specific to the skyline queries/operators.
Fix bug where incomplete skylines may not be properly be calculated in case of cyclic dominance relationships.

### Are there _any_ user-facing changes?

The syntax for skyline queries was changed from

[ COMPLETE ] SKYLINE OF [ DISTINCT ] ...

to

SKYLINE OF [ DISTINCT ] [ COMPLETE ] ...