[SKYLINE][SQL] Add optimization where skyline is pushed through a join if it only involves one side of the join.
An optimization rule is added to the Catalyst optimizer which analyzes the skyline and join. If the skyline only involves one side of the join it is pushed through to that side.

### What changes were proposed?

Introduction of an optimization rule to push skylines through joins.

### Why are the changes needed?

Changes are needed to increase the performance of skyline queries by computing the skyline early and reducing the input size for both the skyline and the join.

### Are there _any_ user-facing changes?

None