[SKYLINE][SQL] Add option for forcing the block-nested-loop algorithm for skyline computation
### What changes were proposed?

Instead of COMPLETE, BNL can now be specified in the skyline algorithm. This forces Spark to use the (non-distributed and complete) block-nested-loop skyline algorithm.

### Why are the changes needed?

The changes are needed to enable better testing and comparisons of algorithms.

### Are there _any_ user-facing changes?

The skyline syntax now supports the following syntax:

... SKYLINE OF [DISTINCT] BNL ...

This also includes the (Scala) DataFrame methods skylineBNL and skylineDistinctBNL which work analogously to the existing skyline methods of the DataFrame API.